+++ ninja-core-demo/src/main/java/controllers/ApplicationController.java
@@ -27,6 +27,8 @@ import ninja.Context;
 import ninja.Result;
 import ninja.Results;
 import ninja.Router;
+import ninja.cache.Cache;
+import ninja.cache.NinjaCache;
 import ninja.i18n.Lang;
 import ninja.i18n.Messages;
 import ninja.params.Param;
@@ -64,7 +66,9 @@ public class ApplicationController {
 
     @Inject
     Router router;
-
+    
+    @Inject
+    NinjaCache ninjaCache;
 
     public Result examples(@Request HttpServletRequest httpServletRequest,
                            @Response HttpServletResponse httpServletResponse,
@@ -196,6 +200,29 @@ public class ApplicationController {
         String maliciousJavascript = "<script>alert('Hello');</script>";
 
         return Results.html().render("maliciousJavascript", maliciousJavascript);
+
+    }
+    
+    
+    public Result testCaching() {
+        
+        // Simple integration test to check if ehcache works:
+        // Calling that route two times will lead to different results.
+        
+        String cacheKeyObject = ninjaCache.get("key", String.class);
+        
+        if (cacheKeyObject != null) {
+            
+            String cacheKey = cacheKeyObject;
+            
+            return Results.html().render("cacheKey", cacheKey);
+            
+        } else {
+            
+            ninjaCache.set("key", "cacheKeyValue", "10s");
+            return Results.html();
+        }
+        
 
     }