+++ ninja-core/src/main/java/ninja/Result.java
@@ -16,8 +16,8 @@
 
 package ninja;
 
+import java.io.Serializable;
 import java.util.AbstractMap;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -28,7 +28,9 @@ import ninja.utils.SwissKnife;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 
-public class Result {
+public class Result implements Serializable {
+
+    private static final long serialVersionUID = -6982988314360788360L;
 
     // /////////////////////////////////////////////////////////////////////////
     // HTTP Status codes (for convenience)