+++ ninja-core/src/main/java/ninja/postoffice/guice/PostofficeProvider.java
@@ -43,6 +43,7 @@ public class PostofficeProvider implements Provider<Postoffice> {
     }
 
     @Override
+       
     public Postoffice get() {
 
         if (mailer == null) {